fix(marathon-gradle-plugin): use platform-dependant script path
